+++ src/app/shared/components/round-button/fs-round-button.component.ts
@@ -1,18 +1,10 @@
-import { Component, OnInit } from '@angular/core';
-import { faChevronRight } from '@fortawesome/free-solid-svg-icons';
+import { Component } from '@angular/core';
 
 @Component({
   selector: 'fs-round-button',
   templateUrl: './fs-round-button.component.html',
-  styleUrls: ['./fs-round-button.component.scss']
+  styleUrls: ['./fs-round-button.component.scss'],
 })
-export class RoundButton implements OnInit {
-
-  chevronIcon = faChevronRight;
-
-  constructor() { }
-
-  ngOnInit(): void {
-  }
-
+export class RoundButton {
+  constructor() {}
 }