+++ Changelog.md
@@ -1,11 +1,24 @@
 # Changelog
 
-
 ## [Unreleased]
 
 ## 2021-05-20/avz
 # Changes   
 - changed mediaquery to hide the list-style-type dots in the footer 
+
+## 2021-05-19/ml
+### Added
+- courseCarousel as a module
+- added Hammerjs
+
+## 2021-05-19/ml
+- script-loader added (dynamically loading javascript-files)
+- KI-Tools are lazy loaded
+
+## 2021-05-18/ml
+- Shared Module (fonts, bootstrap Modules and Shared Components)
+- LandingPage has its own module now
+- Shared Component LoadingSpinner
 
 ## 2021-05-06/ml    
 New Project with Angular 11